REFACTOR: Rename app from 'topics' to 'content_management'
This commit renames the 'topics' app to 'content_management' for clarity within the project. The new name better reflects the purpose of the app which is to manage content-related functionality such as topics, sources, and content items, and doesn't create confusion with the Topic model.

Changes made include updating directory names, file references, and configurations throughout the project to reflect the new app name.