REFACTOR: Update content and source fields in Comment model
This commit modifies the `content` and `source` fields in the Comment model to allow them to be optional (i.e., possibly blank) when creating comments.

Changes:
- Altered the `content` field to allow it to be optional.
- Altered the `source` field to allow it to be optional.

These changes provide flexibility in creating comments without specifying the associated content or source. I'm in the process of figuring out if these two fields are needed in the model.