FIX: Resolve issue with Instagram reels and pagination
This commit addresses an issue where some Instagram reels were not displaying correctly on the page when the short version of the embed code was used. The bug was fixed by adding conditional loading of the Instagram embed script within the if statement of the resource_detail.html. Embed script is now included only when the content source is Instagram, ensuring proper rendering of the reels without affecting the loading of other content.

Additionally, to prevent overloading and improve performance, the pagination for the page was adjusted to display a maximum of 3 contents per page.