DOCS: Add Modules imported in README.md
This commit updates the README.md file by adding details about the modules used in the project. The list of modules includes both standard Python and Django modules, as well as custom modules specific to the project. This addition provides transparency regarding the technology stack used in the project and helps the developer to easily remember the modules used.